requirements.txt
# Core Flask and extensions
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Admin
Flask-Migrate
Flask-Mail

# Database driver
psycopg2-binary

# Payments
stripe

# Background Tasks
celery
redis

# WSGI Server
gunicorn

# AI SDKs
openai

# Utilities
python-dotenv
itsdangerous
click
requests

# Testing
pytest
pytest-flask
pytest-mock

# Monitoring
sentry-sdk[flask]